+++ src/Encamina.Enmarcha.SemanticKernel/MemoryStoreExtender.cs
@@ -1,9 +1,11 @@
 ﻿// Ignore Spelling: Upsert
 
-using System;
 using System.Collections.ObjectModel;
+using System.Net;
 using System.Runtime.CompilerServices;
 using System.Text.Json;
+
+using Azure;
 
 using Encamina.Enmarcha.SemanticKernel.Abstractions;
 using Encamina.Enmarcha.SemanticKernel.Abstractions.Events;
@@ -132,15 +134,30 @@ public class MemoryStoreExtender : IMemoryStoreExtender
     private async Task<int> GetChunkSize(string memoryId, string collectionName, CancellationToken cancellationToken)
     {
         var key = BuildMemoryIdentifier(memoryId, 0);
-        var fistMemoryChunk = await MemoryStore.GetAsync(collectionName, key, cancellationToken: cancellationToken);
+        MemoryRecord firstMemoryChunk = null;
+
+        try
+        {
+            firstMemoryChunk = await MemoryStore.GetAsync(collectionName, key, cancellationToken: cancellationToken);
+        }
+        catch (RequestFailedException e) when (e.Status == (int)HttpStatusCode.NotFound)
+        {
+            // At this point, we need to catch the NotFound exception. This is necessary because, in the case of Azure AI Search Memory Store,
+            // if the element does not exist, it throws an exception instead of returning a null value, which would be the expected behavior.
+            // We have opened an issue in Semantic Kernel and also an issue in ENMARCHA to track the progress of this issue and remove this try/catch block once it is resolved.
+            // Issue ENMARCHA: https://github.com/Encamina/enmarcha/issues/72
+
+            /* Do nothing */
+        }
+
         RaiseMemoryStoreEvent(new() { EventType = MemoryStoreEventTypes.Get, Keys = [key], CollectionName = collectionName });
 
-        if (fistMemoryChunk == null)
+        if (firstMemoryChunk == null)
         {
             return 0;
         }
 
-        var metadata = JsonSerializer.Deserialize<Dictionary<string, string>>(fistMemoryChunk.Metadata.AdditionalMetadata);
+        var metadata = JsonSerializer.Deserialize<Dictionary<string, string>>(firstMemoryChunk.Metadata.AdditionalMetadata);
 
         return int.Parse(metadata[ChunkSize]);
     }